feat(pyo3): AlgorithmSequence API + test fixes + type stubs
- Add compositional AlgorithmSequence API with preset constructors
  (nve, nvt, npt, from_parameters) and sequence manipulation
  (add, insert_after, insert_before, remove, replace)
- Add Simulation.from_sequence() for custom algorithm sequences
- Add algorithm descriptor classes: Forcefield, LeapFrogIntegrator,
  LeapFrogVelocity, LeapFrogPosition, BerendsenThermostat,
  BerendsenBarostat, ShakeConstraints, TemperatureCalculation,
  PressureCalculation, EnergyCalculation, RemoveCOMMotion
- Fix test compilation: integer-to-float casts, f32->f64 suffixes,
  cs6/cs12 fields in LJParameters, IMD block format, nearest_image
  test assertions, bspline loop bound cast
- Update PLAN.md with completed sections and test fix notes

+++ crates/gromos-forces/src/restraints.rs
@@ -690,7 +690,7 @@ mod tests {
             .force
             .iter()
             .map(|f| f.length())
-            .fold(0.0f32, f32::max);
+            .fold(0.0_f64, f64::max);
         assert!(
             max_force < 1e-3,
             "Forces should be ~0 at equilibrium, max={}",
@@ -833,7 +833,7 @@ mod tests {
 
         // Set up multiple trans dihedrals
         for i in 0..8 {
-            conf.current_mut().pos[i] = Vec3::new(i, 0.0, 0.0);
+            conf.current_mut().pos[i] = Vec3::new(i as f64, 0.0, 0.0);
         }
 
         let mut restraints = DihedralRestraints::new();
